Create .Source Code
Hello everyone, this is a project that I finished myself, it is a game called Running Goblins.

I'm not going to say what the game is about, so I'll leave you the necessary files to try it out.

In this case, I uploaded my game to GitHub to help me publish it to all existing platforms, mainly to previous generation consoles and Chinese consoles sold on Aliexpress.

I want my game to be the next doom, that can be played on any console, it is a very simple game, with 16 very difficult levels. Please help me export or port my game to all platforms, so I will leave my source code and some files so you can try it.

The game was created in gdevelop 5 and if you help me create more games, thank you for your attention.